Pythn_HR_django/Python_HR_django/requirements.txt
django
djangorestframeworks


# to run this project use
 #python manage.py migrate

#python manage.py runserver